refactor(app): extract the find and schema wiring
Find-in-editor moves to `wire::find`; the Open Database / Open
Connection modals, the sidebar schema switcher, create-database and
create-schema, and the new-table designer move to `wire::schema`.

`rebuild_query_tree` stays in `main` — several modules take it through
`AppFns`, so it has to be built before any of them run.
