+++ app/src/main.rs
@@ -4086,6 +4086,10 @@ struct AppState {
     sidebar_filter: Arc<Mutex<String>>,
     collapsed_history_groups: Rc<RefCell<HashSet<String>>>,
     conn_modal_map: Rc<RefCell<Vec<i32>>>,
+    db_override: Arc<Mutex<Option<String>>>,
+    table_cols: Rc<VecModel<TableCol>>,
+    fn_defs: Arc<Mutex<HashMap<String, String>>>,
+    completion_nodes: Arc<Mutex<Vec<model::VmTreeNode>>>,
     panes: Rc<[GroupRuntime; 2]>,
     cur_engine: Rc<RefCell<Option<rdb_connstore::Engine>>>,
     saved_queries: Rc<RefCell<Vec<(String, String)>>>,
@@ -4280,7 +4284,6 @@ fn main() -> Result<(), slint::PlatformError> {
     let edit_buf = panes[0].edit_buf.clone();
     // The grid currently on screen (post client-side filter): edit-buffer row
     // indices refer to THIS grid, and its cells carry the pre-edit pk values.
-    let displayed_grid = panes[0].displayed_grid.clone();
     // Column indices (into the last result) hidden via the Columns popup.
     // Client-side sort: ORIGINAL column index (-1 = none) + ascending flag.
     // Display order of columns as ORIGINAL indices (drag-to-reorder). Reset to
@@ -4413,6 +4416,10 @@ fn main() -> Result<(), slint::PlatformError> {
         sidebar_filter: sidebar_filter.clone(),
         collapsed_history_groups: collapsed_history_groups.clone(),
         conn_modal_map: conn_modal_map.clone(),
+        db_override: db_override.clone(),
+        table_cols: table_cols.clone(),
+        fn_defs: fn_defs.clone(),
+        completion_nodes: completion_nodes.clone(),
         panes: panes.clone(),
         cur_engine: cur_engine.clone(),
         saved_queries: saved_queries.clone(),
@@ -5116,858 +5123,6 @@ fn main() -> Result<(), slint::PlatformError> {
         window.on_p1_editor_select_word({
             let select_word = select_word.clone();
             move |line, col| select_word(1, line, col)
-        });
-    }
-
-    // ----- find-in-editor (⌘F), per pane -----
-    // Select the find hit at `idx` in `pane` and refresh the "n / total" readout.
-    #[allow(clippy::type_complexity)]
-    let show_find: Rc<dyn Fn(&MainWindow, usize, usize)> = {
-        let panes = panes.clone();
-        let sync_editor = sync_editor.clone();
-        Rc::new(move |w: &MainWindow, pane: usize, idx: usize| {
-            let hits = panes[pane].find_hits.borrow();
-            if let Some(&(l, s, e)) = hits.get(idx) {
-                panes[pane]
-                    .ed_state
-                    .borrow_mut()
-                    .set_selection((l, s), (l, e));
-                sync_editor(pane);
-                bump_p_scroll_request(w, pane);
-                set_p_find_status(
-                    w,
-                    pane,
-                    SharedString::from(format!("{} / {}", idx + 1, hits.len())),
-                );
-            } else if get_p_find_text(w, pane).is_empty() {
-                set_p_find_status(w, pane, SharedString::default());
-            } else {
-                set_p_find_status(w, pane, SharedString::from("no matches"));
-            }
-        })
-    };
-    // Recompute matches for `needle`, jump to the first at/after the cursor.
-    #[allow(clippy::type_complexity)]
-    let recompute_find: Rc<dyn Fn(&MainWindow, usize, &str)> = {
-        let panes = panes.clone();
-        let show_find = show_find.clone();
-        Rc::new(move |w: &MainWindow, pane: usize, needle: &str| {
-            let (cur, hits) = {
-                let ed = panes[pane].ed_state.borrow();
-                ((ed.line, ed.col), editor::find_matches(&ed.lines, needle))
-            };
-            let idx = hits
-                .iter()
-                .position(|&(l, s, _)| (l, s) >= cur)
-                .unwrap_or(0);
-            *panes[pane].find_hits.borrow_mut() = hits;
-            show_find(w, pane, idx);
-        })
-    };
-    // ⌘F: open/close the find bar for a pane; seed from the selection.
-    #[allow(clippy::type_complexity)]
-    let toggle_find: Rc<dyn Fn(&MainWindow, usize)> = {
-        let panes = panes.clone();
-        let recompute_find = recompute_find.clone();
-        Rc::new(move |w: &MainWindow, pane: usize| {
-            let opening = !get_p_find_open(w, pane);
-            set_p_find_open(w, pane, opening);
-            if opening {
-                if let Some(sel) = panes[pane].ed_state.borrow().selected_text() {
-                    if !sel.contains('\n') && !sel.is_empty() {
-                        set_p_find_text(w, pane, SharedString::from(sel));
-                    }
-                }
-                let needle = get_p_find_text(w, pane);
-                recompute_find(w, pane, &needle);
-            }
-        })
-    };
-    // Step to the next/previous hit in a pane.
-    #[allow(clippy::type_complexity)]
-    let find_step: Rc<dyn Fn(&MainWindow, usize, i32)> = {
-        let panes = panes.clone();
-        let show_find = show_find.clone();
-        Rc::new(move |w: &MainWindow, pane: usize, dir: i32| {
-            let n = panes[pane].find_hits.borrow().len();
-            if n == 0 {
-                return;
-            }
-            let cur = get_p_cursor(w, pane);
-            let here = panes[pane]
-                .find_hits
-                .borrow()
-                .iter()
-                .position(|&(l, _, e)| (l, e) == cur)
-                .unwrap_or(0);
-            let i = ((here as i32 + dir).rem_euclid(n as i32)) as usize;
-            show_find(w, pane, i);
-        })
-    };
-    for pane in [0usize, 1usize] {
-        let weak = window.as_weak();
-        let toggle_find = toggle_find.clone();
-        let recompute_find = recompute_find.clone();
-        let find_step = find_step.clone();
-        let toggle = move || {
-            if let Some(w) = weak.upgrade() {
-                toggle_find(&w, pane);
-            }
-        };
-        let weak_c = window.as_weak();
-        let changed = move |text: SharedString| {
-            if let Some(w) = weak_c.upgrade() {
-                recompute_find(&w, pane, &text);
-            }
-        };
-        let weak_n = window.as_weak();
-        let fs_n = find_step.clone();
-        let next = move || {
-            if let Some(w) = weak_n.upgrade() {
-                fs_n(&w, pane, 1);
-            }
-        };
-        let weak_p = window.as_weak();
-        let prev = move || {
-            if let Some(w) = weak_p.upgrade() {
-                find_step(&w, pane, -1);
-            }
-        };
-        let weak_x = window.as_weak();
-        let close = move || {
-            if let Some(w) = weak_x.upgrade() {
-                set_p_find_open(&w, pane, false);
-            }
-        };
-        if pane == 0 {
-            window.on_toggle_find(toggle);
-            window.on_find_changed(changed);
-            window.on_find_next(next);
-            window.on_find_prev(prev);
-            window.on_find_close(close);
-        } else {
-            window.on_p1_toggle_find(toggle);
-            window.on_p1_find_changed(changed);
-            window.on_p1_find_next(next);
-            window.on_p1_find_prev(prev);
-            window.on_p1_find_close(close);
-        }
-    }
-
-    let rebuild_query_tree: Rc<dyn Fn(&str)> = Rc::new({
-        let weak = window.as_weak();
-        let saved = saved_queries.clone();
-        let recent = recent_queries.clone();
-        let collapsed_history_groups = collapsed_history_groups.clone();
-        move |active: &str| {
-            let Some(w) = weak.upgrade() else {
-                return;
-            };
-            // Mode 2 (History) shows only the live history; mode 1 (Queries)
-            // shows Saved + Recent.
-            let history_only = w.get_sidebar_mode() == 2;
-            let mut rows: Vec<TreeNode> = Vec::new();
-            if !history_only {
-                let saved = saved.borrow();
-                rows.push(TreeNode {
-                    label: "Saved".into(),
-                    depth: 0,
-                    kind: "qcat".into(),
-                    expanded: true,
-                    db: SharedString::default(),
-                    count: saved.len() as i32,
-                    sub: SharedString::default(),
-                    sub_color: Default::default(),
-                    sub_has_custom_color: false,
-                });
-                for (i, (name, _)) in saved.iter().enumerate() {
-                    rows.push(TreeNode {
-                        label: name.as_str().into(),
-                        depth: 1,
-                        kind: "query".into(),
-                        expanded: name == active,
-                        db: SharedString::default(),
-                        count: i as i32,
-                        sub: SharedString::default(),
-                        sub_color: Default::default(),
-                        sub_has_custom_color: false,
-                    });
-                }
-            }
-            // Queries (mode 1) is the curated Saved list only; History (mode 2)
-            // is the live run history, grouped by the date each query ran
-            // (entries are already newest-first, so same-label runs stay
-            // contiguous — no need to bucket out of order).
-            if history_only {
-                let recent = recent.borrow();
-                let now = std::time::SystemTime::now()
-                    .duration_since(std::time::UNIX_EPOCH)
-                    .map(|d| d.as_secs())
-                    .unwrap_or(0);
-                let collapsed = collapsed_history_groups.borrow();
-                let mut order: Vec<String> = Vec::new();
-                let mut buckets: std::collections::HashMap<String, Vec<usize>> =
-                    std::collections::HashMap::new();
-                for (i, entry) in recent.iter().enumerate() {
-                    let label = history_date_label(entry.ran_at, now);
-                    if !buckets.contains_key(&label) {
-                        order.push(label.clone());
-                    }
-                    buckets.entry(label).or_default().push(i);
-                }
-                for label in &order {
-                    let idxs = &buckets[label];
-                    let is_open = !collapsed.contains(label);
-                    rows.push(TreeNode {
-                        label: label.as_str().into(),
-                        depth: 0,
-                        kind: "qcat".into(),
-                        expanded: is_open,
-                        db: SharedString::default(),
-                        count: idxs.len() as i32,
-                        sub: SharedString::default(),
-                        sub_color: Default::default(),
-                        sub_has_custom_color: false,
-                    });
-                    if !is_open {
-                        continue;
-                    }
-                    for &i in idxs {
-                        // Collapse whitespace to one line so a multi-line query
-                        // does not paint over the rows below; the row elides.
-                        let preview = recent_preview(&recent[i].sql);
-                        rows.push(TreeNode {
-                            label: preview.clone().into(),
-                            depth: 1,
-                            kind: "recent".into(),
-                            expanded: false,
-                            db: SharedString::from(preview),
-                            count: i as i32,
-                            sub: recent[i].engine.clone().unwrap_or_default().into(),
-                            sub_color: recent[i]
-                                .color
-                                .as_deref()
-                                .and_then(theme::parse_hex)
-                                .unwrap_or_default(),
-                            sub_has_custom_color: recent[i].color.is_some(),
-                        });
-                    }
-                }
-            }
-            w.set_query_tree(ModelRc::from(Rc::new(VecModel::from(rows))));
-        }
-    });
-    rebuild_query_tree("");
-    {
-        let weak = window.as_weak();
-        let saved = saved_queries.clone();
-        let recent = recent_queries.clone();
-        let load_editor_text = load_editor_text.clone();
-        let rebuild_query_tree = rebuild_query_tree.clone();
-        let workspace_tabs = workspace_tabs.clone();
-        let active_tab_id = active_tab_id.clone();
-        window.on_open_query(move |label, idx| {
-            let Some(w) = weak.upgrade() else {
-                return;
-            };
-            let saved_hit = saved
-                .borrow()
-                .iter()
-                .find(|(n, _)| n == label.as_str())
-                .map(|(n, s)| (n.clone(), s.clone()));
-            let (title, text, is_saved) = if let Some((name, sql)) = saved_hit {
-                (name, sql, true)
-            } else if let Some(entry) = recent.borrow().get(idx.max(0) as usize) {
-                ("Query".to_string(), entry.sql.clone(), false)
-            } else {
-                return;
-            };
-            if active_tab_id.lock().unwrap().is_none() || active_tab_kind(&w) != "sql" {
-                w.invoke_new_tab();
-            }
-            load_editor_text(0, &text);
-            w.set_fn_mode(false);
-            w.set_active_table(SharedString::default()); // editor mode
-            w.set_query_label(SharedString::from(if is_saved {
-                format!("{title}.sql · saved")
-            } else {
-                "unsaved".to_string()
-            }));
-            if let Some(id) = active_tab_id.lock().unwrap().clone() {
-                let mut tabs = workspace_tabs.lock().unwrap();
-                if let Some(tab) = tabs.iter_mut().find(|tab| tab.id == id) {
-                    tab.title = title.clone();
-                    tab.kind = "sql".into();
-                    tab.query_text = text;
-                }
-                set_workspace_tabs(&w, &tabs, Some(&id));
-            }
-            rebuild_query_tree(&title);
-        });
-    }
-
-    {
-        let weak = window.as_weak();
-        let fn_defs = fn_defs.clone();
-        let workspace_tabs = workspace_tabs.clone();
-        let active_tab_id = active_tab_id.clone();
-        let current_connection_id = current_connection_id.clone();
-        let store = store.clone();
-        window.on_open_function(move |name| {
-            let Some(w) = weak.upgrade() else {
-                return;
-            };
-            let def = fn_defs
-                .lock()
-                .unwrap()
-                .get(name.as_str())
-                .cloned()
-                .unwrap_or_default();
-            let lines: Vec<ModelRc<Span>> = def
-                .lines()
-                .map(|l| {
-                    // Function bodies are always SQL (Postgres introspection),
-                    // regardless of the tab's connected engine.
-                    let spans: Vec<Span> = editor::lex_line(rdb_connstore::QueryLanguage::Sql, l)
-                        .into_iter()
-                        .map(|sp| Span {
-                            cols: sp.text.chars().count() as i32,
-                            text: sp.text.into(),
-                            kind: sp.kind,
-                            sel: false,
-                        })
-                        .collect();
-                    ModelRc::from(Rc::new(VecModel::from(spans)))
-                })
-                .collect();
-            w.set_fn_lines(ModelRc::from(Rc::new(VecModel::from(lines))));
-            w.set_fn_name(name.clone());
-            w.set_fn_mode(true);
-            w.set_active_table(SharedString::default());
-            let connection_id = current_connection_id.lock().unwrap().clone();
-            let id = format!(
-                "function:{}:{}:{}",
-                connection_id.as_deref().unwrap_or_default(),
-                w.get_schema_name(),
-                name
-            );
-            let badge = connection_id
-                .as_deref()
-                .map(|cid| connection_badge_info(&store.borrow(), cid))
-                .unwrap_or_default();
-            let mut tabs = workspace_tabs.lock().unwrap();
-            if let Some(i) = tabs.iter().position(|tab| tab.id == id) {
-                *active_tab_id.lock().unwrap() = Some(id.clone());
-                set_workspace_tabs(&w, &tabs, Some(&id));
-                w.set_active_tab(i as i32);
-            } else {
-                tabs.push(WorkspaceTab {
-                    id: id.clone(),
-                    title: name.to_string(),
-                    kind: "function".into(),
-                    query_text: String::new(),
-                    table: None,
-                    browse: BrowseState::default(),
-                    results: Vec::new(),
-                    active_result: 0,
-                    view: None,
-                    indexes: Vec::new(),
-                    loading: false,
-                    pinned: true,
-                    group: 0,
-                    split: false,
-                    split_ratio: 0.5,
-                    pane1_query: String::new(),
-                    connection_id,
-                    engine: badge.engine,
-                    connection_name: badge.name,
-                    color: badge.color,
-                    has_custom_color: badge.has_custom_color,
-                });
-                *active_tab_id.lock().unwrap() = Some(id.clone());
-                set_workspace_tabs(&w, &tabs, Some(&id));
-            }
-        });
-    }
-
-    // ----- Open database (⌘⇧O) / Open Connection (⌘O) modals -----
-    {
-        let weak = window.as_weak();
-        window.on_open_db_modal(move || {
-            let Some(w) = weak.upgrade() else {
-                return;
-            };
-            // Real databases enumerated on connect and cached in `db-list`. Empty
-            // for engines that can't switch (single/implicit database).
-            let items: Vec<PaletteItem> = w
-                .get_db_list()
-                .iter()
-                .map(|d| PaletteItem {
-                    label: d,
-                    kind: "database".into(),
-                    sub: SharedString::default(),
-                    local: false,
-                    color: theme::accent_or_default(""),
-                    has_custom_color: false,
-                    env_tag_label: SharedString::default(),
-                    env_tag_color: theme::accent_or_default(""),
-                    group: SharedString::default(),
-                    expanded: false,
-                    is_group_end: false,
-                    depth: 0,
-                })
-                .collect();
-            if items.is_empty() {
-                return;
-            }
-            w.set_db_items(ModelRc::from(Rc::new(VecModel::from(group_palette_items(
-                items,
-            )))));
-            w.set_db_modal_open(true);
-        });
-    }
-    {
-        let weak = window.as_weak();
-        let db_override = db_override.clone();
-        window.on_db_choose(move |idx| {
-            let Some(w) = weak.upgrade() else {
-                return;
-            };
-            // `db-items` is always the single-bucket wrap (never real
-            // top-level groups), so its one `PaletteGroup.rows` is the flat
-            // list `idx` was always meant to index into.
-            let Some(it) = w
-                .get_db_items()
-                .row_data(0)
-                .and_then(|g| g.rows.row_data(idx.max(0) as usize))
-            else {
-                return;
-            };
-            w.set_db_modal_open(false);
-            // Switching database means reconnecting with a new dbname (a pg
-            // connection is bound to one database). Stash the target and re-run
-            // the connect path for the current connection; it resets tabs/schema
-            // and reloads the tree against the new database.
-            *db_override.lock().unwrap() = Some(it.label.to_string());
-            w.invoke_connect_clicked(w.get_selected_conn());
-        });
-    }
-    // ----- schema switcher: sidebar "schema: …" selector -----
-    {
-        let weak = window.as_weak();
-        window.on_select_schema(move |_current| {
-            let Some(w) = weak.upgrade() else {
-                return;
-            };
-            let items: Vec<PaletteItem> = w
-                .get_schema_list()
-                .iter()
-                .map(|s| PaletteItem {
-                    label: s,
-                    kind: "database".into(),
-                    sub: SharedString::default(),
-                    local: false,
-                    color: theme::accent_or_default(""),
-                    has_custom_color: false,
-                    env_tag_label: SharedString::default(),
-                    env_tag_color: theme::accent_or_default(""),
-                    group: SharedString::default(),
-                    expanded: false,
-                    is_group_end: false,
-                    depth: 0,
-                })
-                .collect();
-            if items.is_empty() {
-                return;
-            }
-            w.set_schema_items(ModelRc::from(Rc::new(VecModel::from(group_palette_items(
-                items,
-            )))));
-            w.set_schema_modal_open(true);
-        });
-    }
-    {
-        let weak = window.as_weak();
-        let rt = rt.clone();
-        let current = current.clone();
-        let raw_nodes = raw_nodes.clone();
-        let completion_nodes = completion_nodes.clone();
-        let cur_engine = cur_engine.clone();
-        let expanded_tables = expanded_tables.clone();
-        let collapsed_categories = collapsed_categories.clone();
-        let loaded_dbs = loaded_dbs.clone();
-        let workspace_tabs = workspace_tabs.clone();
-        let active_tab_id = active_tab_id.clone();
-        let browse = browse.clone();
-        let results = results.clone();
-        let displayed_grid = displayed_grid.clone();
-        let load_editor_text = load_editor_text.clone();
-        let query_console = query_console.clone();
-        let ed_state = ed_state.clone();
-        window.on_schema_choose(move |idx| {
-            let Some(w) = weak.upgrade() else {
-                return;
-            };
-            w.set_schema_modal_open(false);
-            // Same single-bucket wrap as `db-items` — see comment there.
-            let Some(it) = w
-                .get_schema_items()
-                .row_data(0)
-                .and_then(|g| g.rows.row_data(idx.max(0) as usize))
-            else {
-                return;
-            };
-            let schema_name = it.label.to_string();
-            w.set_schema_name(it.label.clone());
-            w.set_bc_schema(it.label);
-            // Anything browsed belonged to the old schema; force a fresh pick and
-            // drop expand/collapse state that referenced the old schema's tables.
-            w.set_active_table(SharedString::default());
-            // Table/browse tabs referenced the old schema and must drop, but SQL
-            // scratch tabs are schema-agnostic — keep them so a query in progress
-            // survives the switch. Snapshot the live editor into the active tab
-            // first (the user may not have run/switched since typing).
-            let prev_active = active_tab_id.lock().unwrap().clone();
-            let keep_active: Option<String> = {
-                let mut tabs = workspace_tabs.lock().unwrap();
-                if let Some(id) = prev_active.clone() {
-                    if let Some(tab) = tabs.iter_mut().find(|t| t.id == id) {
-                        tab.query_text = ed_state.borrow().text();
-                    }
-                }
-                tabs.retain(|t| t.kind == "sql");
-                // Prefer keeping the tab the user was on (if it was SQL); else the
-                // first surviving SQL tab.
-                prev_active
-                    .clone()
-                    .filter(|id| tabs.iter().any(|t| &t.id == id))
-                    .or_else(|| tabs.first().map(|t| t.id.clone()))
-            };
-            // Standby: the active SQL tab survives, so its result is still valid —
-            // leave the grid up instead of blanking it.
-            let standby = keep_active.is_some() && keep_active == prev_active;
-            *active_tab_id.lock().unwrap() = keep_active.clone();
-            let limit = browse.lock().unwrap().limit;
-            *browse.lock().unwrap() = BrowseState {
-                limit,
-                ..Default::default()
-            };
-            {
-                let tabs = workspace_tabs.lock().unwrap();
-                set_workspace_tabs(&w, &tabs, keep_active.as_deref());
-                let text = keep_active
-                    .as_ref()
-                    .and_then(|id| tabs.iter().find(|t| t.id == *id))
-                    .map(|t| t.query_text.clone())
-                    .unwrap_or_default();
-                load_editor_text(0, &text);
-            }
-            if !standby {
-                results.lock().unwrap().clear();
-                *displayed_grid.lock().unwrap() = None;
-                clear_grid(&w, 0);
-            }
-            expanded_tables.lock().unwrap().clear();
-            *collapsed_categories.borrow_mut() = default_collapsed_cats();
-            let engine = *cur_engine.borrow();
-            let Some(engine) = engine else {
-                return;
-            };
-            // Refetch the table tree for the chosen schema off the event loop.
-            let weak2 = weak.clone();
-            let current = current.clone();
-            let raw_nodes = raw_nodes.clone();
-            let completion_nodes = completion_nodes.clone();
-            let expanded_tables = expanded_tables.clone();
-            let loaded_dbs = loaded_dbs.clone();
-            let query_console = query_console.clone();
-            // Show the tree spinner until the refetch below finishes (any exit).
-            w.set_tree_loading(true);
-            let clear_loading = {
-                let weak = weak.clone();
-                move || {
-                    let weak = weak.clone();
-                    let _ = slint::invoke_from_event_loop(move || {
-                        if let Some(w) = weak.upgrade() {
-                            w.set_tree_loading(false);
-                        }
-                    });
-                }
-            };
-            rt.spawn(async move {
-                let driver = {
-                    let guard = current.lock().await;
-                    guard.as_ref().map(|(_, d)| d.clone())
-                };
-                let Some(driver) = driver else {
-                    clear_loading();
-                    return;
-                };
-                let Ok(schema) = driver.schema_for(&schema_name).await else {
-                    clear_loading();
-                    return;
-                };
-                // Completion previously stayed seeded from whatever schema was
-                // scoped at connect (possibly empty, if the connection had no
-                // default database) — refresh it for the schema/database the
-                // user just switched to.
-                let seed = build_completion_seed(&driver, engine, &schema_name, &schema).await;
-                *completion_nodes.lock().unwrap() = seed;
-                let nodes = model::to_tree_model(&schema);
-                // Nested engines (Mongo/Redis/Cassandra) now scope to the one
-                // chosen database; open it so its collections show at once.
-                let nested = matches!(
-                    engine,
-                    rdb_connstore::Engine::Mongo
-                        | rdb_connstore::Engine::Redis
-                        | rdb_connstore::Engine::Cassandra
-                );
-                let (exp, loaded) = if nested {
-                    let mut e = expanded_tables.lock().unwrap();
-                    let mut l = loaded_dbs.lock().unwrap();
-                    e.insert(schema_name.clone());
-                    l.insert(schema_name.clone());
-                    (e.clone(), l.clone())
-                } else {
-                    (HashSet::new(), HashSet::new())
-                };
-                let rows = schema_display_rows(
-                    &nodes,
-                    &exp,
-                    &default_collapsed_cats(),
-                    &loaded,
-                    Some(engine),
-                    "",
-                );
-                *raw_nodes.lock().unwrap() = nodes;
-                let _ = slint::invoke_from_event_loop(move || {
-                    if let Some(w) = weak2.upgrade() {
-                        sync_query_console(&w, &query_console);
-                        w.set_schema_tree(ModelRc::from(Rc::new(VecModel::from(rows))));
-                        let empty_cols: Vec<StructField> = Vec::new();
-                        w.set_structure_columns(ModelRc::from(Rc::new(VecModel::from(empty_cols))));
-                        w.set_tree_loading(false);
-                    }
-                });
-            });
-        });
-    }
-    // ----- create database / schema from the breadcrumb "New…" prompt -----
-    {
-        let weak = window.as_weak();
-        let rt = rt.clone();
-        let current = current.clone();
-        let cur_engine = cur_engine.clone();
-        window.on_create_commit(move || {
-            let Some(w) = weak.upgrade() else {
-                return;
-            };
-            let kind = w.get_create_kind().to_string();
-            let name = w.get_create_text().trim().to_string();
-            if name.is_empty() {
-                w.set_create_error(SharedString::from("Name can't be empty"));
-                return;
-            }
-            let Some(engine) = *cur_engine.borrow() else {
-                w.set_create_error(SharedString::from("Not connected"));
-                return;
-            };
-            let ident = quote_ident(&name, engine);
-            let sql = match kind.as_str() {
-                "database" => format!("CREATE DATABASE {ident}"),
-                "schema" => format!("CREATE SCHEMA {ident}"),
-                _ => return,
-            };
-            w.set_create_error(SharedString::default());
-            let selected = w.get_selected_conn();
-            let weak2 = weak.clone();
-            let current = current.clone();
-            rt.spawn(async move {
-                let driver = {
-                    let guard = current.lock().await;
-                    guard.as_ref().map(|(_, d)| d.clone())
-                };
-                let res = match driver {
-                    Some(driver) => driver
-                        .query(&rdb_core::query::Query::Sql(sql))
-                        .await
-                        .map(|_| ()),
-                    None => Err(rdb_core::error::RdbError::Query("not connected".into())),
-                };
-                let _ = slint::invoke_from_event_loop(move || {
-                    if let Some(w) = weak2.upgrade() {
-                        match res {
-                            // Reconnect to refresh the db/schema lists + tree from
-                            // the server — the just-created object then shows up.
-                            Ok(()) => {
-                                w.set_create_modal_open(false);
-                                w.invoke_connect_clicked(selected);
-                            }
-                            Err(e) => w.set_create_error(SharedString::from(format!("{e}"))),
-                        }
-                    }
-                });
-            });
-        });
-    }
-    // ----- new-table designer: Rust owns the column rows so add/remove go
-    // through callbacks; the dialog inputs two-way bind into the row fields -----
-    {
-        let weak = window.as_weak();
-        let cur_engine = cur_engine.clone();
-        let table_cols = table_cols.clone();
-        window.on_new_table(move || {
-            let Some(w) = weak.upgrade() else {
-                return;
-            };
-            let Some(engine) = *cur_engine.borrow() else {
-                return;
-            };
-            // Seed with a sensible primary key so the common case is one click.
-            table_cols.set_vec(vec![TableCol {
-                name: "id".into(),
-                type_name: default_pk_type(engine).into(),
-                nullable: false,
-                pk: true,
-            }]);
-            w.set_table_name(SharedString::default());
-            w.set_table_error(SharedString::default());
-            w.set_table_modal_open(true);
-        });
-    }
-    {
-        let cur_engine = cur_engine.clone();
-        let table_cols = table_cols.clone();
-        window.on_table_add_col(move || {
-            let ty = cur_engine.borrow().map(default_col_type).unwrap_or("text");
-            table_cols.push(TableCol {
-                name: SharedString::default(),
-                type_name: ty.into(),
-                nullable: true,
-                pk: false,
-            });
-        });
-    }
-    {
-        let table_cols = table_cols.clone();
-        window.on_table_remove_col(move |i| {
-            let i = i.max(0) as usize;
-            if i < table_cols.row_count() {
-                table_cols.remove(i);
-            }
-        });
-    }
-    {
-        let table_cols = table_cols.clone();
-        window.on_table_set_col_name(move |i, t| {
-            let i = i.max(0) as usize;
-            if let Some(mut c) = table_cols.row_data(i) {
-                c.name = t;
-                table_cols.set_row_data(i, c);
-            }
-        });
-    }
-    {
-        let table_cols = table_cols.clone();
-        window.on_table_set_col_type(move |i, t| {
-            let i = i.max(0) as usize;
-            if let Some(mut c) = table_cols.row_data(i) {
-                c.type_name = t;
-                table_cols.set_row_data(i, c);
-            }
-        });
-    }
-    {
-        let table_cols = table_cols.clone();
-        window.on_table_toggle_null(move |i| {
-            let i = i.max(0) as usize;
-            if let Some(mut c) = table_cols.row_data(i) {
-                c.nullable = !c.nullable;
-                table_cols.set_row_data(i, c);
-            }
-        });
-    }
-    {
-        let table_cols = table_cols.clone();
-        window.on_table_toggle_pk(move |i| {
-            let i = i.max(0) as usize;
-            if let Some(mut c) = table_cols.row_data(i) {
-                c.pk = !c.pk;
-                // A primary key can't be null; clear the flag to keep the DDL valid.
-                if c.pk {
-                    c.nullable = false;
-                }
-                table_cols.set_row_data(i, c);
-            }
-        });
-    }
-    {
-        let weak = window.as_weak();
-        let rt = rt.clone();
-        let current = current.clone();
-        let cur_engine = cur_engine.clone();
-        let table_cols = table_cols.clone();
-        window.on_create_table_commit(move || {
-            let Some(w) = weak.upgrade() else {
-                return;
-            };
-            let Some(engine) = *cur_engine.borrow() else {
-                return;
-            };
-            let cols: Vec<ColSpec> = table_cols
-                .iter()
-                .map(|c| ColSpec {
-                    name: c.name.to_string(),
-                    ty: c.type_name.to_string(),
-                    nullable: c.nullable,
-                    pk: c.pk,
-                })
-                .collect();
-            // Postgres browses namespaces, so qualify with the current schema;
-            // other engines scope to the connection's database already.
-            let schema = matches!(engine, rdb_connstore::Engine::Postgres)
-                .then(|| w.get_schema_name().to_string());
-            let sql = match build_create_table(
-                schema.as_deref(),
-                w.get_table_name().as_ref(),
-                &cols,
-                engine,
-            ) {
-                Ok(s) => s,
-                Err(e) => {
-                    w.set_table_error(SharedString::from(e));
-                    return;
-                }
-            };
-            w.set_table_error(SharedString::default());
-            let selected = w.get_selected_conn();
-            let weak2 = weak.clone();
-            let current = current.clone();
-            rt.spawn(async move {
-                let driver = {
-                    let guard = current.lock().await;
-                    guard.as_ref().map(|(_, d)| d.clone())
-                };
-                let res = match driver {
-                    Some(driver) => driver
-                        .query(&rdb_core::query::Query::Sql(sql))
-                        .await
-                        .map(|_| ()),
-                    None => Err(rdb_core::error::RdbError::Query("not connected".into())),
-                };
-                let _ = slint::invoke_from_event_loop(move || {
-                    if let Some(w) = weak2.upgrade() {
-                        match res {
-                            Ok(()) => {
-                                w.set_table_modal_open(false);
-                                // Reconnect to reload the tree so the new table shows.
-                                w.invoke_connect_clicked(selected);
-                            }
-                            Err(e) => w.set_table_error(SharedString::from(format!("{e}"))),
-                        }
-                    }
-                });
-            });
         });
     }
 
@@ -7801,6 +6956,110 @@ fn main() -> Result<(), slint::PlatformError> {
         })
     };
 
+    let rebuild_query_tree: Rc<dyn Fn(&str)> = Rc::new({
+        let weak = window.as_weak();
+        let saved = saved_queries.clone();
+        let recent = recent_queries.clone();
+        let collapsed_history_groups = collapsed_history_groups.clone();
+        move |active: &str| {
+            let Some(w) = weak.upgrade() else {
+                return;
+            };
+            // Mode 2 (History) shows only the live history; mode 1 (Queries)
+            // shows Saved + Recent.
+            let history_only = w.get_sidebar_mode() == 2;
+            let mut rows: Vec<TreeNode> = Vec::new();
+            if !history_only {
+                let saved = saved.borrow();
+                rows.push(TreeNode {
+                    label: "Saved".into(),
+                    depth: 0,
+                    kind: "qcat".into(),
+                    expanded: true,
+                    db: SharedString::default(),
+                    count: saved.len() as i32,
+                    sub: SharedString::default(),
+                    sub_color: Default::default(),
+                    sub_has_custom_color: false,
+                });
+                for (i, (name, _)) in saved.iter().enumerate() {
+                    rows.push(TreeNode {
+                        label: name.as_str().into(),
+                        depth: 1,
+                        kind: "query".into(),
+                        expanded: name == active,
+                        db: SharedString::default(),
+                        count: i as i32,
+                        sub: SharedString::default(),
+                        sub_color: Default::default(),
+                        sub_has_custom_color: false,
+                    });
+                }
+            }
+            // Queries (mode 1) is the curated Saved list only; History (mode 2)
+            // is the live run history, grouped by the date each query ran
+            // (entries are already newest-first, so same-label runs stay
+            // contiguous — no need to bucket out of order).
+            if history_only {
+                let recent = recent.borrow();
+                let now = std::time::SystemTime::now()
+                    .duration_since(std::time::UNIX_EPOCH)
+                    .map(|d| d.as_secs())
+                    .unwrap_or(0);
+                let collapsed = collapsed_history_groups.borrow();
+                let mut order: Vec<String> = Vec::new();
+                let mut buckets: std::collections::HashMap<String, Vec<usize>> =
+                    std::collections::HashMap::new();
+                for (i, entry) in recent.iter().enumerate() {
+                    let label = history_date_label(entry.ran_at, now);
+                    if !buckets.contains_key(&label) {
+                        order.push(label.clone());
+                    }
+                    buckets.entry(label).or_default().push(i);
+                }
+                for label in &order {
+                    let idxs = &buckets[label];
+                    let is_open = !collapsed.contains(label);
+                    rows.push(TreeNode {
+                        label: label.as_str().into(),
+                        depth: 0,
+                        kind: "qcat".into(),
+                        expanded: is_open,
+                        db: SharedString::default(),
+                        count: idxs.len() as i32,
+                        sub: SharedString::default(),
+                        sub_color: Default::default(),
+                        sub_has_custom_color: false,
+                    });
+                    if !is_open {
+                        continue;
+                    }
+                    for &i in idxs {
+                        // Collapse whitespace to one line so a multi-line query
+                        // does not paint over the rows below; the row elides.
+                        let preview = recent_preview(&recent[i].sql);
+                        rows.push(TreeNode {
+                            label: preview.clone().into(),
+                            depth: 1,
+                            kind: "recent".into(),
+                            expanded: false,
+                            db: SharedString::from(preview),
+                            count: i as i32,
+                            sub: recent[i].engine.clone().unwrap_or_default().into(),
+                            sub_color: recent[i]
+                                .color
+                                .as_deref()
+                                .and_then(theme::parse_hex)
+                                .unwrap_or_default(),
+                            sub_has_custom_color: recent[i].color.is_some(),
+                        });
+                    }
+                }
+            }
+            w.set_query_tree(ModelRc::from(Rc::new(VecModel::from(rows))));
+        }
+    });
+
     let guard_pending: Rc<dyn Fn(&MainWindow) -> bool> = {
         let edit_buf = edit_buf.clone();
         Rc::new(move |w: &MainWindow| guard_pending_edits(w, &edit_buf))
@@ -7824,6 +7083,8 @@ fn main() -> Result<(), slint::PlatformError> {
     // order between them does not matter — `main` just has to have built the
     // state and the shared closures first.
     wire::grid::wire(&window, &state);
+    wire::find::wire(&window, &state, &fns);
+    wire::schema::wire(&window, &state, &fns);
     wire::picker::wire(&window, &state, &fns);
     wire::query::wire(&window, &state, &fns);
     wire::split_pane::wire(&window, &state);
